Split files into its own headers and impls
